requirements.txt
bcrypt==3.2.0
boto3==1.16.14
cryptography==3.2.1
Flask==1.1.2
Flask-Caching==1.9.0
Flask-Cors==3.0.9
Flask-JWT-Extended==3.24.1
Flask-SQLAlchemy==2.4.4
future==0.18.2
mysql-connector-python==8.0.22
passlib==1.7.2
PyJWT==1.7.1
PyMySQL==0.10.1
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
simplejson==3.17.2
SQLAlchemy==1.3.20
Werkzeug==1.0.1
python-dotenv==0.19.2
wsgi==0.20
